requirements.txt
selenium==3.141.0
validators==0.15.0
discord.py==1.4.1
tldextract==2.2.2
python-dotenv==0.13.0
aiohttp==3.6.2
fake-useragent==0.1.11
lxml==4.5.1
pysqlcipher3==1.0.3
pytz==2020.1
schedule~=0.6.0